backend/requirements.txt
# Core Flask framework
Flask==3.0.3

# Flask extensions
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# Database driver (PostgreSQL)
psycopg2-binary==2.9.9

# Alembic (used internally by Flask-Migrate)
alembic==1.13.3

# Security & authentication
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
email-validator==2.1.1

# Background tasks
celery==5.4.0
redis==5.1.0
